Project/requirements.txt
asgiref==3.4.1
cycler==0.11.0
Django==3.2.25
django-dbbackup==4.0.2
djangorestframework==3.15.1
importlib-resources==5.4.0
kiwisolver==1.3.1
matplotlib
numpy
Pillow==8.4.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
tkinter-page==0.0.5
typing-extensions==4.1.1
zipp==3.6.0
environs==9.5.0
gunicorn==21.2.0